reformatting test-shadows.sh and alter debug code

+++ test-shadows.sh
@@ -5,7 +5,7 @@
 # add-shadows.bsh
 
 
-:;: 'Regular users only, and -sudo- required' :;:
+: 'Regular users only, and -sudo- required' 
 if [[ "${UID}" == 0 ]]; then
   printf '\n\t Must be a regular user and use sudo. \n\n'
   exit "${LINENO}"
@@ -15,29 +15,28 @@ elif ! sudo -v; then
   exit "${LINENO}"
 fi
 
-:;: 'Target string:' :;:
-set -vx
+: 'Target string:' 
+#set -vx
 if [[ "$#" -eq 0 ]]; then x='export'; else x="$1"; fi;
 readonly x
-declare -p x
+#declare -p x
 enable -n "$x"
-enable -a | grep "$x"
+#enable -a | grep "$x"
 #exit "${LINENO}"
 
-:;: 'Required programs' ;:;
+: 'Required programs' ;:;
 if [[ "${BASH_VERSION:0:1}" -lt 5 ]]; then 
   echo Please install Bash version 5, thanks.
   exit "${LINENO}"
 fi
 
-reqd_cmds=( awk chmod cp cut dirname find grep ln ls 
-  mkdir rm rmdir stat sudo tee )
+reqd_cmds=( awk chmod cp cut dirname find grep ln ls mkdir rm rmdir 
+  stat sudo tee )
 yn=n
 
 hash -r; 
-hash
+#hash
 for c in "${reqd_cmds[@]}"; do 
-  hash -p "$(type -P sudo)" sudo
   if type -P "${c}" > /dev/null 2>&1; then 
     hash -p "$(type -P "${c}")" "${c}"
   else
@@ -45,7 +44,7 @@ for c in "${reqd_cmds[@]}"; do
     list+=("${c}")
   fi; 
 done; unset c reqd_cmds
-hash
+#hash
 
 if [[ "${yn}" == 'n' ]]; then 
   : 'No additional commands are required'
@@ -56,89 +55,99 @@ else
   exit "${LINENO}"
 fi; unset yn list
 #exit "${LINENO}"
-#set -x
+#set -vx
 
 
-:;: 'Functions, variables and umask' :;:
+: 'Functions, variables and umask' 
 function fn_erx(){
   local ec="${?}"
   echo ERROR: "${@}"
   exit "${ec}"
 }
 
+: 'Set up testing of PATH dirs' :
 unset PATH 
 PATH="/home/liveuser/.local/bin_symlink"  # < 0
 PATH+=":/home/liveuser/bin_hardlink"      # < 1
 PATH+=":/usr/local/bin_copy-of-inode"     # < 2
-PATH+=":/usr/bin"
+PATH+=":/usr/bin"                         # < 3
 PATH+=":/bin"
 PATH+=":/usr/local/sbin_dangling_symlink" # < 5
 PATH+=":/usr/sbin"
-declare -p PATH
-
-: 'Set up testing of PATH dirs' :
+#declare -p PATH
 IFS=':' read -ra find_path <<< "${PATH}"
-declare -p find_path
+#declare -p find_path
 
+: 'remove any pre-existing shadow-files for name-string "$x"'
 for dirnm in "${find_path[@]}"; do
   fullnm="${dirnm}/${x}"
   if [[ -f "${fullnm}" ]] || [[ -L "${fullnm}" ]]
   then
     sudo rm -f --one-file-system --preserve-root=all -- "${fullnm}" || 
       fn_erx
-  fi; unset fullnm; :
-done; unset dirnm; :
+  fi; unset fullnm
+done; unset dirnm
 
-: 'Set up testing of files' :
-xfa="/usr/bin/${x}"
-xfb="${find_path[0]}/${x}" # symlink
-xfc="${find_path[1]}/${x}" # hardlink
-xfd="${find_path[2]}/${x}" # copy of inode
-xfe="${find_path[5]}/${x}" # dangling symlink
+: 'Set up testing of shadow-files' :
+xfa="${find_path[0]}/${x}" # symlink              # < 0
+xfb="${find_path[1]}/${x}" # hardlink             # < 1
+xfc="${find_path[2]}/${x}" # copy of inode        # < 2
+xfd="/usr/bin/${x}"        # executable file      # < 3
+xfe="${find_path[5]}/${x}" # dangling symlink     # < 5
 xff=("${xfa}" "${xfb}" "${xfc}" "${xfd}" "${xfe}")
 umask 022
 #exit "${LINENO}"
 set -x
 
 
-:;: 'Tests: Remove certain values' :;:
+: 'Tests: Remove certain values' 
 
-:;: 'Remove: Regular variable' :;:
+: 'Remove: Regular variable' 
 unset -v "${x}"
 
-:;: 'Remove: Namerefs' :;:
+: 'Remove: Namerefs' 
 unset -n "${x}"
 
 # This code block: for testing only
-:;: 'Remove: Unused dirs in PATH' :;:
+: 'Remove: Unused dirs in PATH' 
 for f in "${xff[@]}"; do
-  d="$(dirname "${f}")"
+  
+  # get dirname
+  d="${f%/*}"
+  
+  # if the dir exists...
   if [[ -d "${d}" ]]; then
+
+    # find any FSO's wi it, but only the first 32 bytes thereof...
     lo="$(sudo find "${d}" -mindepth 1 -printf '%p' | head -c 32)"
   fi
-  if [[ -z "${lo:0:32}" ]]; then
-    if [[ -d "${d}" ]] && [[ ! -L "${d}" ]]
-    then
-      sudo rmdir -- "${d}" || 
-        fn_erx "${LINENO}"
-    fi
-  fi; unset d lo; :
-done; unset f; :
 
-:;: 'Remove: Builtin' :;:
+  # 
+  if [[ -v lo ]]; then   
+    if [[ -z "${lo:0:32}" ]]; then
+      if [[ -d "${d}" ]] && [[ ! -L "${d}" ]]
+      then
+        sudo rmdir -- "${d}" || 
+          fn_erx "${LINENO}"
+      fi
+    fi; unset d lo
+  fi; unset d lo
+done; unset f
+
+: 'Remove: Builtin' 
 enable -n "${x}"
 
-:;: 'Remove: Function' :;:
+: 'Remove: Function' 
 unset -f "${x}"
 
-:;: 'Remove: Alias' :;:
+: 'Remove: Alias' 
 unalias -a 
 exit "${LINENO}"
 #set -x
 
 
 
-:;: 'Create Shadow variable' :;:
+: 'Create Shadow variable' 
 dpo="$(declare -p "${x}" 2> /dev/null)"
 if [[ -z "${dpo:0:8}" ]]; then
   eval "${x}"=quux
@@ -152,7 +161,7 @@ fi; unset dpo
 
 
 
-:;: 'Create Shadow nameref' :;:
+: 'Create Shadow nameref' 
 dao="$(declare -p "${x}" | awk '$2 ~ /n/')"
 if [[ -z "${dao:0:8}" ]]; then
   declare -n "${x}"=USER
@@ -166,20 +175,20 @@ fi; unset dao
 
 
 
-:;: 'create PATH dirs as necc' :;:
+: 'create PATH dirs as necc' 
 for f in "${xff[@]}"; do
   d="$(dirname "${f}")"
   if [[ ! -d "${d}" ]]; then
     sudo mkdir -p "${d}" || 
       fn_erx "${LINENO}"
-  fi; unset d; :
-done; unset f; :
+  fi; unset d
+done; unset f
 #exit "${LINENO}"
 #set -x
 
 
 
-:;: 'Create Shadow executable file' :;:
+: 'Create Shadow executable file' 
 if [[ ! -f "${xfa}" ]]; then
   printf '\x23\x21/usr/bin/sh\n%s \x22\x24\x40\x22\n' "${x}" | 
     sudo tee  "${xfa}" > /dev/null || 
@@ -196,7 +205,7 @@ fi
 
 
 
-:;: 'Symlink of shadow file' :;:
+: 'Symlink of shadow file' 
 if [[ ! -f "${xfb}" ]]; then
   sudo ln -s "${xfa}" "${xfb}" || 
     fn_erx "${LINENO}"
@@ -212,7 +221,7 @@ fi
 
 
 
-:;: 'Hardlink of shadow file' :;:
+: 'Hardlink of shadow file' 
 if [[ ! -f "${xfc}" ]]; then
   sudo ln "${xfa}" "${xfc}" || 
     fn_erx "${LINENO}"
@@ -228,7 +237,7 @@ fi
 
 
 
-:;: 'Dangling symlink of shadow file' :;:
+: 'Dangling symlink of shadow file' 
 if [[ ! -f "${xfd}" ]]; then
   sudo cp -b "${xfa}" "${xfd}" || 
     fn_erx "${LINENO}"
@@ -251,14 +260,14 @@ fi
 if [[ "$(type -t "${x}")" != file ]]; then
   fn_erx "${LINENO}"
 fi
-:;: 'the actual file has been removed so the extra symlink may "dangle"' :;:
+: 'the actual file has been removed so the extra symlink may "dangle"' 
 unset 'xff[3]'
 #exit "${LINENO}"
 #set -x
 
 
 
-:;: 'DAC permissions of shadow files' :;:
+: 'DAC permissions of shadow files' 
 for f in "${xff[@]}" ; do
   if [[ -L "${f}" ]]; then
     continue
@@ -272,14 +281,14 @@ for f in "${xff[@]}" ; do
     if ! grep -q 755 <<< "${so}"; then
       fn_erx
     fi
-  fi; :
-done; :
+  fi
+done
 #exit "${LINENO}"
 #set -x
 
 
 
-:;: 'Builtin' :;:
+: 'Builtin' 
 eago="$(enable | grep "${x}")"
 if [[ -z "${eago:0:8}" ]]; then
   enable "${x}"
@@ -296,7 +305,7 @@ fi
 
 
 
-:;: 'Create Shadow function' :;:
+: 'Create Shadow function' 
 dpfo="$(declare -pf "${x}" 2> /dev/null)"
 if [[ -z "${dpfo:0:8}" ]]; then
   : 'define function'
@@ -314,7 +323,7 @@ fi
 
 
 
-:;: 'Create Shadow alias' :;:
+: 'Create Shadow alias' 
 shopt -s expand_aliases
 ao="$(alias "${x}" 2> /dev/null)"
 if [[ -z "${ao:0:8}" ]]; then
@@ -332,7 +341,7 @@ fi
 
 
 
-:;: 'Verification 1: type -P' :;:
+: 'Verification 1: type -P' 
 printf '\n\t hash -r; hash; type -P \n\n' 
 hash -r; hash; type -P "${x}" 
 type_P_o="$(type -P "${x}")"
@@ -344,7 +353,7 @@ ls -alhFiL --quoting-style=shell-always --color=always "${type_P_o}"
 
 
 
-:;: 'Verification 2: type -a' :;:
+: 'Verification 2: type -a' 
 printf '\n\t declare -p find_path # (Same as PATH.) \n\n' 
 declare -p find_path
 printf '\n\t command -pV; command -v \n\n'
